apigateway_listener refactoring (#1048)
* Extract all helper methods from apigateway_listener

* Remove overriding the apigateway return_response
- createdDate is no longer string and was fixed in moto (https://github.com/spulec/moto/pull/1457)
