+++ localstack/services/apigateway/helpers.py
@@ -1,5 +1,131 @@
 import re
+import json
+
+from localstack.utils import common
+from requests.models import Response
 from localstack.utils.aws import aws_stack
+from six.moves.urllib import parse as urlparse
+from localstack.constants import APPLICATION_JSON
+
+# regex path patterns
+PATH_REGEX_MAIN = r'^/restapis/([A-Za-z0-9_\-]+)/[a-z]+(\?.*)?'
+PATH_REGEX_SUB = r'^/restapis/([A-Za-z0-9_\-]+)/[a-z]+/([A-Za-z0-9_\-]+)/.*'
+
+# maps API ids to authorizers
+AUTHORIZERS = {}
+
+
+def _create_response_object(content, code, headers):
+    response = Response()
+    response.status_code = code
+    response.headers = headers
+    response._content = content
+    return response
+
+
+def make_response(message):
+    return _create_response_object(json.dumps(message), code=200, headers={'Content-Type': APPLICATION_JSON})
+
+
+def flask_to_requests_response(r):
+    return _create_response_object(r.data, code=r.status_code, headers=r.headers)
+
+
+def make_error(message, code=400):
+    response = Response()
+    response.status_code = code
+    response._content = json.dumps({'message': message})
+    return response
+
+
+def get_api_id_from_path(path):
+    match = re.match(PATH_REGEX_SUB, path)
+    if match:
+        return match.group(1)
+    return re.match(PATH_REGEX_MAIN, path).group(1)
+
+
+def get_authorizers(path):
+    result = {'item': []}
+    api_id = get_api_id_from_path(path)
+    for key, value in AUTHORIZERS.items():
+        auth_api_id = get_api_id_from_path(value['_links']['self']['href'])
+        if auth_api_id == api_id:
+            result['item'].append(value)
+    return result
+
+
+def add_authorizer(path, data):
+    api_id = get_api_id_from_path(path)
+    result = common.clone(data)
+    result['id'] = common.short_uid()
+    if '_links' not in result:
+        result['_links'] = {}
+    result['_links']['self'] = {
+        'href': '/restapis/%s/authorizers/%s' % (api_id, result['id'])
+    }
+    AUTHORIZERS[result['id']] = result
+    return result
+
+
+def handle_authorizers(method, path, data, headers):
+    result = {}
+    if method == 'GET':
+        result = get_authorizers(path)
+    elif method == 'POST':
+        result = add_authorizer(path, data)
+    else:
+        return make_error('Not implemented for API Gateway authorizers: %s' % method, 404)
+    return make_response(result)
+
+
+def tokenize_path(path):
+    return path.lstrip('/').split('/')
+
+
+def extract_path_params(path, extracted_path):
+    tokenized_extracted_path = tokenize_path(extracted_path)
+    # Looks for '{' in the tokenized extracted path
+    path_params_list = [(i, v) for i, v in enumerate(tokenized_extracted_path) if '{' in v]
+    tokenized_path = tokenize_path(path)
+    path_params = {}
+    for param in path_params_list:
+        path_param_name = param[1][1:-1].encode('utf-8')
+        path_param_position = param[0]
+        if path_param_name.endswith(b'+'):
+            path_params[path_param_name] = '/'.join(tokenized_path[path_param_position:])
+        else:
+            path_params[path_param_name] = tokenized_path[path_param_position]
+    path_params = common.json_safe(path_params)
+    return path_params
+
+
+def extract_query_string_params(path):
+    parsed_path = urlparse.urlparse(path)
+    path = parsed_path.path
+    parsed_query_string_params = urlparse.parse_qs(parsed_path.query)
+
+    query_string_params = {}
+    for query_param_name, query_param_values in parsed_query_string_params.items():
+        if len(query_param_values) == 1:
+            query_string_params[query_param_name] = query_param_values[0]
+        else:
+            query_string_params[query_param_name] = query_param_values
+
+    return [path, query_string_params]
+
+
+def get_cors_response(headers):
+    # TODO: for now we simply return "allow-all" CORS headers, but in the future
+    # we should implement custom headers for CORS rules, as supported by API Gateway:
+    # http://docs.aws.amazon.com/apigateway/latest/developerguide/how-to-cors.html
+    response = Response()
+    response.status_code = 200
+    response.headers['Access-Control-Allow-Origin'] = '*'
+    response.headers['Access-Control-Allow-Methods'] = 'GET, POST, PUT, DELETE'
+    response.headers['Access-Control-Allow-Headers'] = '*'
+    response._content = ''
+    return response
 
 
 def get_rest_api_paths(rest_api_id):
